action authentication with cli
